combined bouqeutsviewcontroller and bouqetControlelr
